Enable Hazelcast when running Arquillian tests
Template for the domain.xml is from the default domain used by Payara
when running in the embedded mode, it has Hazelcast enabled + some
stuff is removed. I'm sure there is a better way to do it.

It's also a bit tricky with relative path in the  Arquillian container configuration.
src/test/resources/domain.xml works from Maven command line, but it require
to set working directory when executed from IDE.